Move policies artifacts
