+++ tests/artifacts/policies/aloha_sim_insertion_human_act_/grad_stats.safetensors
