+++ tests/artifacts/policies/aloha_sim_insertion_human_act_1000_steps/output_dict.safetensors
