+++ tests/artifacts/policies/pusht_diffusion_/grad_stats.safetensors
