+++ tests/artifacts/policies/pusht_diffusion_/param_stats.safetensors
