+++ tests/artifacts/policies/xarm_lift_medium_tdmpc_use_mpc/actions.safetensors
