+++ tests/artifacts/policies/xarm_lift_medium_tdmpc_use_mpc/grad_stats.safetensors
