+++ tests/artifacts/policies/xarm_lift_medium_tdmpc_use_mpc/param_stats.safetensors
