+++ tests/artifacts/policies/xarm_lift_medium_tdmpc_use_policy/output_dict.safetensors
